Fixed: Albums queries are now as fast as in XBMC.
git-svn-id: http://android-xbmcremote.googlecode.com/svn/trunk/XBMC Remote@218 677bec2c-7919-11de-bdc7-1f637afd43b1

+++ src/org/xbmc/android/backend/httpapi/HttpApiHandler.java
@@ -38,13 +38,13 @@ import android.app.Activity;
 public class HttpApiHandler<T> implements Runnable {
 	public T value;
 	protected final Activity mActivity;
-	protected final int mTag;
+	protected final long mTag;
 	protected CacheType mCacheType;
 	public HttpApiHandler(Activity activity) {
 		mActivity = activity;
 		mTag = 0;
 	}
-	public HttpApiHandler(Activity activity, int tag) {
+	public HttpApiHandler(Activity activity, long tag) {
 		mActivity = activity;
 		mTag = tag;
 	}
