+++ src/org/xbmc/httpapi/data/ICoverArt.java
@@ -2,6 +2,6 @@ package org.xbmc.httpapi.data;
 
 public interface ICoverArt {
 	public String getArtFolder();
-	public int getCrc();
+	public long getCrc();
 	public int getFallbackCrc();
 }
